backend/requirements.txt
# Django
Django==2.1.5
django-cors-headers==2.4.0
django-environ==0.4.4
django-extensions==2.0.7
pytest-django==3.2.1


# Django REST Framework
djangorestframework==3.8.2'